Random Song Play
This enables the theme's ability to pick a random song from your SM Songs library and plays it as the menu's music.

Not the best implementation in the world, but it's ok I guess.

You're also able to select the folder to play the songs on or the option to play all of them. When it's enabled, when entering Music Selection, the MusicWheel cursor will go to the song that was just playing.