Update Blast (that was dumb)
- Implemented Character Selection screen
(It's still a work in progress. Feedback would be really appreciated. Only Player one that use this screen at the moment)

- Incremented the size of the Difficulty list in ScreenSelectMusic
(It was quite small after play-testing the theme on smaller screen resolutions.)

- Cleanup in metrics.ini
(Removed some unnecesary commands that I just don't need anymore.)

- Camera fix for Japonica
(Camera 5 was being problematic because of the z position it was being sent by the Default_Camera.)