Add Camera Debug
This adds a new screen that allows you to control the camera to whatever you like to go.

It allows for XYZ movement, as well as rotation. Both modes can be toggled by the Start button. To control the Z axis set of buttons, you'll need to have some input set to the EffectUp/Down buttons.