Experimental Light Fix
This experimental fix, reduces the brightness of the Character during gameplay. This can be useful to fix model material colors from being completely whited out by the max diffuse that is used by default.

Keep in mind that some models that rely more on texture lighting might give graphical issues if this setting is enabled.